Move API call to utils, add degree of parallelism.
